Add files via upload
The Interactive Quiz Website is a web-based platform designed to provide  users with a seamless and engaging experience for taking quizzes on a variety  of topics. Built using modern web development technologies, the platform  allows users to select from a range of quiz categories, answer multiple-choice  questions, and receive immediate feedback on their performance. The website  is designed to be responsive, ensuring compatibility across diﬀerent devices,  including desktops, tablets, and smartphones.
The project uses a combination of HTML, CSS, and JavaScript for the frontend,  providing an intuitive and interactive user interface.